Add boolean email_sent/sms_sent columns to message
Add email_sent/sms_sent columns to message table with default FALSE.
These will be used to monitor notifications sent asynchronously, with
the async function changing the default value to TRUE/1 when the
operation is completed.
A celery/chron job can be regularly run to check for any recently unsent
notifications.
Datetime-ing each (message rec'd, notifications sent) might be a useful
addition when this is implemented, to check only for recently unsent
notifications, although if the unsent queue is emptied regularly, this
would not be necessary, aside from performance monitoring.