Add target="_blank" rel="noreferrer" to ext links.
This involved placing a blog_url attribute on the application object,
and a context processor inside the application factory returning the
attribute to supply the url to both the jinja template and to the blog
route from the a single point of truth (were the url to change). This
allows the /blog route to remain functional, with an automatic redirect.
This could just as easily be placed in a config variable, but since it's
 a link/content more than it is application configuration, it is placed
 as code in to the repository.

A refactor to the nav bar links is made to enable links (ie blog) to
external sites to be rendered without url_for{internal.route).